+++ src/contentsharecontroller/DefaultContentShareController.ts
@@ -11,6 +11,7 @@ import MeetingSessionCredentials from '../meetingsession/MeetingSessionCredentia
 import MeetingSessionStatus from '../meetingsession/MeetingSessionStatus';
 import DefaultModality from '../modality/DefaultModality';
 import AsyncScheduler from '../scheduler/AsyncScheduler';
+import VideoCodecCapability from '../sdp/VideoCodecCapability';
 import { Maybe } from '../utils/Types';
 import VideoTile from '../videotile/VideoTile';
 import ContentShareSimulcastEncodingParameters from '../videouplinkbandwidthpolicy/ContentShareSimulcastEncodingParameters';
@@ -109,6 +110,10 @@ export default class DefaultContentShareController
         Maybe.of(observer.contentShareDidUnpause).map(f => f.call(observer));
       });
     }
+  }
+
+  setContentShareVideoCodecPreferences(preferences: VideoCodecCapability[]): void {
+    this.contentAudioVideo.setVideoCodecSendPreferences(preferences);
   }
 
   async destroy(): Promise<void> {